--- src/Microsoft.Performance.SDK.Runtime/ColumnBuilding/IDataColumnVariantsProcessor.cs
@@ -1,9 +1,0 @@
-using Microsoft.Performance.SDK.Runtime.ColumnVariants;
-
-namespace Microsoft.Performance.SDK.Runtime.ColumnBuilding
-{
-    public interface IDataColumnVariantsProcessor
-    {
-        void ProcessColumnVariants(IColumnVariant variants);
-    }
-}
