+++ src/Microsoft.Performance.SDK.Runtime/ColumnVariants/ModesToggleColumnVariant.cs
@@ -1,47 +1,59 @@
-namespace Microsoft.Performance.SDK.Runtime.ColumnVariants
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.Performance.SDK.Runtime.ColumnVariants;
+
+/// <summary>
+///     Exposes a toggle for enabling/disabling a <see cref="IColumnVariant"/>.
+///     This <see cref="IColumnVariant"/> has no associated projection.
+/// </summary>
+public sealed class ModesToggleColumnVariant
+    : IColumnVariant
 {
-    /// <summary>
-    ///     Represents a toggle for enabling/disabling <see cref="ModesColumnVariant"/>.
-    ///     This <see cref="IColumnVariant"/> has no associated projection.
-    /// </summary>
-    public sealed class ModesToggleColumnVariant
-        : IColumnVariant
+    public ModesToggleColumnVariant(string toggleText, IColumnVariant modes)
     {
-        public ModesToggleColumnVariant(string toggleText, IColumnVariant modes)
-        {
-            ToggleText = toggleText;
-            Modes = modes;
-        }
+        ToggleText = toggleText;
+        Modes = modes;
+    }
 
-        public string ToggleText { get; }
+    /// <summary>
+    ///     Gets the text to display for the toggle.
+    /// </summary>
+    public string ToggleText { get; }
 
-        public IColumnVariant Modes { get; }
+    /// <summary>
+    ///     Gets the <see cref="IColumnVariant"/> to use when the toggle is enabled.
+    /// </summary>
+    public IColumnVariant Modes { get; }
 
-        public void Accept(IColumnVariantsVisitor visitor)
-        {
-            visitor.Visit(this);
-        }
+    /// <inheritdoc/>
+    public void Accept(IColumnVariantsVisitor visitor)
+    {
+        visitor.Visit(this);
+    }
 
-        private bool Equals(ModesToggleColumnVariant other)
-        {
-            return ToggleText == other.ToggleText && Equals(Modes, other.Modes);
-        }
+    private bool Equals(ModesToggleColumnVariant other)
+    {
+        return ToggleText == other.ToggleText && Equals(Modes, other.Modes);
+    }
 
-        public bool Equals(IColumnVariant other)
-        {
-            return ReferenceEquals(this, other) || other is ModesToggleColumnVariant otherT && Equals(otherT);
-        }
+    /// <inheritdoc/>
+    public bool Equals(IColumnVariant other)
+    {
+        return ReferenceEquals(this, other) || other is ModesToggleColumnVariant otherT && Equals(otherT);
+    }
 
-        public override bool Equals(object obj)
-        {
-            return ReferenceEquals(this, obj) || obj is IColumnVariant other && Equals(other);
-        }
+    /// <inheritdoc/>
+    public override bool Equals(object obj)
+    {
+        return ReferenceEquals(this, obj) || obj is IColumnVariant other && Equals(other);
+    }
 
-        public override int GetHashCode()
-        {
-            return HashCodeUtils.CombineHashCodeValues(
-                this.ToggleText?.GetHashCode() ?? 0,
-                this.Modes?.GetHashCode() ?? 0);
-        }
+    /// <inheritdoc/>
+    public override int GetHashCode()
+    {
+        return HashCodeUtils.CombineHashCodeValues(
+            this.ToggleText?.GetHashCode() ?? 0,
+            this.Modes?.GetHashCode() ?? 0);
     }
 }
