+++ src/Microsoft.Performance.SDK.Runtime/ColumnVariants/NullColumnVariant.cs
@@ -1,5 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 namespace Microsoft.Performance.SDK.Runtime.ColumnVariants;
 
+/// <summary>
+///     Represents a null column variant.
+/// </summary>
 public sealed class NullColumnVariant
     : IColumnVariant
 {
@@ -7,28 +13,35 @@ public sealed class NullColumnVariant
     {
     }
 
-    public static NullColumnVariant Instance { get; } = NullColumnVariant.Instance;
+    /// <summary>
+    ///     Gets the singleton instance of <see cref="NullColumnVariant"/>.
+    /// </summary>
+    public static NullColumnVariant Instance { get; } = new();
 
     private bool Equals(NullColumnVariant other)
     {
         return true;
     }
 
+    /// <inheritdoc/>
     public override bool Equals(object obj)
     {
         return ReferenceEquals(this, obj) || obj is NullColumnVariant other && Equals(other);
     }
 
+    /// <inheritdoc/>
     public override int GetHashCode()
     {
         return 42;
     }
 
+    /// <inheritdoc/>
     public bool Equals(IColumnVariant other)
     {
         return ReferenceEquals(this, other) || other is NullColumnVariant otherT && Equals(otherT);
     }
 
+    /// <inheritdoc/>
     public void Accept(IColumnVariantsVisitor visitor)
     {
         visitor.Visit(this);
