+++ src/Microsoft.Performance.SDK.Runtime/ColumnVariants/ToggleColumnVariant.cs
@@ -7,35 +7,35 @@ using Microsoft.Performance.SDK.Processing.ColumnBuilding;
 namespace Microsoft.Performance.SDK.Runtime.ColumnVariants;
 
 /// <summary>
-///     Represents a column variant that is one of many mutually exclusive modes.
+///     Represents a column variant that can be toggled on or off.
 /// </summary>
-public sealed class ModeColumnVariant
+public sealed class ToggleableColumnVariant
     : IColumnVariant
 {
-    public ModeColumnVariant(
-        ColumnVariantIdentifier modeIdentifier,
-        IDataColumn modeColumn,
+    public ToggleableColumnVariant(
+        ColumnVariantIdentifier identifier,
+        IDataColumn toggledColumn,
         IColumnVariant subVariant)
     {
-        ModeIdentifier = modeIdentifier;
-        ModeColumn = modeColumn;
+        Identifier = identifier;
         SubVariant = subVariant;
+        ToggledColumn = toggledColumn;
     }
 
     /// <summary>
-    ///     Gets the identifier for the mode.
+    ///     Gets the identifier for the column variant.
     /// </summary>
-    public ColumnVariantIdentifier ModeIdentifier { get; }
-
-    /// <summary>
-    ///     Gets the column that this mode represents.
-    /// </summary>
-    public IDataColumn ModeColumn { get; }
+    public ColumnVariantIdentifier Identifier { get; }
 
     /// <summary>
     ///     Gets the sub-variant that represents nested variants of this mode.
     /// </summary>
     public IColumnVariant SubVariant { get; }
+
+    /// <summary>
+    ///     Gets the column that the toggled on state represents.
+    /// </summary>
+    public IDataColumn ToggledColumn { get; }
 
     /// <inheritdoc/>
     public void Accept(IColumnVariantsVisitor visitor)
@@ -43,16 +43,16 @@ public sealed class ModeColumnVariant
         visitor.Visit(this);
     }
 
-    private bool Equals(ModeColumnVariant other)
+    private bool Equals(ToggleableColumnVariant other)
     {
-        return this.ModeIdentifier.Equals(other.ModeIdentifier)
+        return this.Identifier.Equals(other.Identifier)
                && Equals(SubVariant, other.SubVariant);
     }
 
     /// <inheritdoc/>
     public bool Equals(IColumnVariant other)
     {
-        return ReferenceEquals(this, other) || other is ModeColumnVariant otherT && Equals(otherT);
+        return ReferenceEquals(this, other) || other is ToggleableColumnVariant otherT && Equals(otherT);
     }
 
     /// <inheritdoc/>
@@ -65,7 +65,7 @@ public sealed class ModeColumnVariant
     public override int GetHashCode()
     {
         return HashCodeUtils.CombineHashCodeValues(
-            this.ModeIdentifier?.GetHashCode() ?? 0,
+            this.Identifier?.GetHashCode() ?? 0,
             this.SubVariant?.GetHashCode() ?? 0);
     }
 }
