+++ src/Microsoft.Performance.SDK/IsExternalInit.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+// See https://stackoverflow.com/a/64749403
+namespace System.Runtime.CompilerServices
+{
+    internal static class IsExternalInit {}
+}
