+++ src/Microsoft.Performance.SDK/Processing/ColumnBuilding/ColumnVariantIdentifier.cs
@@ -3,35 +3,15 @@
 
 using System;
 
-namespace Microsoft.Performance.SDK.Processing.ColumnBuilding
-{
-    public sealed class ColumnVariantIdentifier
-    {
-        public Guid Id { get; }
+namespace Microsoft.Performance.SDK.Processing.ColumnBuilding;
 
-        private bool Equals(ColumnVariantIdentifier other)
-        {
-            return Id.Equals(other.Id) && Name == other.Name;
-        }
-
-        public override bool Equals(object obj)
-        {
-            return ReferenceEquals(this, obj) || obj is ColumnVariantIdentifier other && Equals(other);
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCodeUtils.CombineHashCodeValues(
-                this.Id.GetHashCode(),
-                this.Name.GetHashCode());
-        }
-
-        public string Name { get; }
-
-        public ColumnVariantIdentifier(Guid id, string name)
-        {
-            this.Id = id;
-            this.Name = name;
-        }
-    }
-}
+/// <summary>
+///     Represents a identifier for a column variant.
+/// </summary>
+/// <param name="Guid">
+///     The unique identifier for the column variant.
+/// </param>
+/// <param name="Name">
+///     The human-readable name of the column variant.
+/// </param>
+public record ColumnVariantIdentifier(Guid Guid, string Name);
