--- src/Microsoft.Performance.SDK/Processing/ColumnBuilding/IColumnVariantsBuilder.cs
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace Microsoft.Performance.SDK.Processing.ColumnBuilding
-{
-    public interface IColumnVariantsBuilder
-    {
-        void Build();
-    }
-}
