+++ src/Microsoft.Performance.SDK/Processing/TableBuilderExtensions.cs
@@ -52,7 +52,7 @@ namespace Microsoft.Performance.SDK.Processing
             this ITableBuilderWithRowCount self,
             ColumnConfiguration column,
             IProjection<int, T> projection,
-            Action<IRootColumnVariantsBuilder> variantsBuilder)
+            Action<IRootColumnBuilder> variantsBuilder)
         {
             Guard.NotNull(self, nameof(self));
             Guard.NotNull(projection, nameof(projection));
